Update authority id

+++ core/primitives/common.hpp
@@ -6,10 +6,12 @@
 #ifndef KAGOME_CORE_PRIMITIVES_COMMON_HPP
 #define KAGOME_CORE_PRIMITIVES_COMMON_HPP
 
+#include <cstdint>
+
 namespace kagome::primitives {
   using BlockNumber = uint64_t;
 
-  using AuthorityId = uint64_t;
-}
+  using AuthorityId = uint32_t;
+}  // namespace kagome::primitives
 
 #endif  // KAGOME_CORE_PRIMITIVES_COMMON_HPP
